fixing code dir structure
