--- DiMo2d/code/code/dipha-2d-thresh/externals/psort-1.0/driver/Make.inc
@@ -1,4 +1,0 @@
-CXX=mpic++
-CXXFLAGS=-O2 -DUSE_FUNNEL -Wall 
-LIBS=-lm
-
