--- DiMo2d/code/code/dipha-2d-thresh/externals/psort-1.0/driver/Makefile
@@ -1,9 +1,0 @@
-include ./Make.inc
-
-psort: psort_driver.cc
-	$(CXX) $(CXXFLAGS) -I../src psort_driver.cc -o psort $(LIBS)
-	strip psort
-
-clean:
-	rm -f *.o psort results
-
