+++ example.py
@@ -18,7 +18,7 @@ def test_2d_func():
     DiMo2d.compute_persistence_single_channel(likelihood_dir, morse_dir, threads, bit_depth, background_pixel_val)
     DiMo2d.generate_morse_graphs(morse_dir, binary_dir, ve_thresh, et_thresh, threads)
     DiMo2d.postprocess_graphs(morse_dir, ve_thresh, et_thresh, threads)
-    DiMo2d.cshl_post_results(morse_dir, json_dir, ve_thresh, et_thresh, threads)
+    # DiMo2d.cshl_post_results(morse_dir, json_dir, ve_thresh, et_thresh, threads)
 
 
 if __name__ == '__main__':
